prime factor

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -4,11 +4,9 @@
  * main - prints largest prime factor.
  * Return: Always 0.
 */
-
 int main(void)
 {
-long int n, fp;
-
+int n, fp;
 n = 612852475143;
 for (fp = 2; fp <= n; fp++)
 {
